third_party/WebKit/LayoutTests/typedcssom/inlinestyle/inlineStylePropertyMap_iteration.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<div id="testElement"></div>

<script>

test(function() {
  testElement.style = "";

  var entryIterator = testElement.attributeStyleMap.entries();
  var keyIterator = testElement.attributeStyleMap.keys();
  var valueIterator = testElement.attributeStyleMap.values();

  assert_true(entryIterator.next().done);
  assert_true(keyIterator.next().done);
  assert_true(valueIterator.next().done);
}, "Iteration over empty StyleMap returns empty iterator");

test(function() {
  testElement.style = "width: 50px";

  var iterator = testElement.attributeStyleMap.entries();
  var entry = iterator.next();
  assert_false(entry.done);
  // Should only be one entry.
  assert_true(iterator.next().done);

  assert_equals(entry.value[0], 'width');
  assert_equals(entry.value[1].constructor.name, CSSUnitValue.name);
  assert_equals(entry.value[1].toString(), '50px');
}, "Iterator for single entry returns iterator with a single value");

test(function() {
  testElement.style = "width: 60px";

  var iterator = testElement.attributeStyleMap.keys();
  var entry = iterator.next();
  assert_false(entry.done);
  // Should only be one entry.
  assert_true(iterator.next().done);

  assert_equals(entry.value, 'width');
}, "Iterator for single key returns iterator with a single value");

test(function() {
  testElement.style = "width: 70px";

  var iterator = testElement.attributeStyleMap.values();
  var entry = iterator.next();
  assert_false(entry.done);
  // Should only be one entry.
  assert_true(iterator.next().done);

  assert_equals(entry.value.constructor.name, CSSUnitValue.name);
  assert_equals(entry.value.toString(), '70px');
}, "Iterator for single value returns iterator with a single value");

test(function() {
  testElement.style = "border: 5px solid lightcoral";

  var entries = {};
  var numEntries = 0;
  for (let value of testElement.attributeStyleMap.entries()) {
    numEntries++;
    entries[value[0]] = value[1];
  }
  assert_equals(numEntries, 17);

  assert_equals(entries['border-top-width'].toString(), '5px');
  assert_equals(entries['border-right-width'].toString(), '5px');
  assert_equals(entries['border-bottom-width'].toString(), '5px');
  assert_equals(entries['border-left-width'].toString(), '5px');

  assert_equals(entries['border-top-style'].toString(), 'solid');
  assert_equals(entries['border-right-style'].toString(), 'solid');
  assert_equals(entries['border-bottom-style'].toString(), 'solid');
  assert_equals(entries['border-left-style'].toString(), 'solid');

  assert_equals(entries['border-top-color'].toString(), 'lightcoral');
  assert_equals(entries['border-right-color'].toString(), 'lightcoral');
  assert_equals(entries['border-bottom-color'].toString(), 'lightcoral');
  assert_equals(entries['border-left-color'].toString(), 'lightcoral');

  assert_equals(entries['border-image-source'].toString(), 'initial');
  assert_equals(entries['border-image-slice'].toString(), 'initial');
  assert_equals(entries['border-image-width'].toString(), 'initial');
  assert_equals(entries['border-image-outset'].toString(), 'initial');
  assert_equals(entries['border-image-repeat'].toString(), 'initial');
}, "Iterating entries over border element expansion");

test(function() {
  testElement.style = "background-color: var(--bg-color);";
  var entries = [...testElement.attributeStyleMap.values()];

  assert_equals(entries.length, 1);
  assert_equals(entries[0].constructor, CSSUnparsedValue);
  assert_equals(entries[0].toString(), 'var(--bg-color)');
  assert_equals([...entries[0]].length, 1);
  assert_equals([...entries[0]][0].variable, "--bg-color");
  assert_equals([...entries[0]][0].fallback, null);
}, "Variable values come out as CSSUnparsedValues");

test(function() {
  testElement.style = '';
  testElement.style.setProperty('--my-custom-property', '5px');

  var entries = [...testElement.attributeStyleMap.entries()];
  assert_equals(entries.length, 1);
  var propertyAndValue = entries[0];
  assert_equals(propertyAndValue.length, 2);

  assert_equals(propertyAndValue[0], '--my-custom-property');
  assert_equals(propertyAndValue[1].constructor, CSSStyleValue);
  assert_equals(propertyAndValue[1].toString(), '5px');
}, "Custom properties set on the element come out as CSSStyleValues");

</script>
